+++ src/api/bookmarkIdolApi.ts
@@ -3,7 +3,6 @@ import { type AxiosResponse } from 'axios';
 import type {
   BookmarkGroup,
   BookmarkIdol,
-  BookmarkSchedule,
   PaginatedResponse,
 } from '@/types/bookmark';
 
@@ -77,23 +76,4 @@ export const addBookmarkIdol = async (idolId: number) => {
  */
 export const removeBookmarkIdol = async (bookmarkId: number) => {
   await axiosInstance.delete(`/bookmarks/idols/${bookmarkId}/`);
-};
-
-/**
- * 내가 북마크한 스케줄 목록 전체 조회
- */
-export const getBookmarkSchedules = async () => {
-  let allSchedules: BookmarkSchedule[] = [];
-  let url: string | null = '/schedules/my/';
-
-  while (url) {
-    // eslint-disable-next-line no-await-in-loop
-    const response: AxiosResponse = await axiosInstance.get<
-      PaginatedResponse<BookmarkSchedule>
-    >(url!);
-    allSchedules = allSchedules.concat(response.data.results);
-    url = response.data.next;
-  }
-
-  return allSchedules;
 };
